some relocation of plugin registry yaml
